+++ src/main/java/com/naribackend/api/ApiControllerAdvice.java
@@ -1,4 +1,4 @@
-package com.naribackend.api.auth;
+package com.naribackend.api;
 
 import com.naribackend.support.error.CoreException;
 import com.naribackend.support.error.ErrorMessage;
