+++ src/main/java/com/naribackend/api/v1/auth/response/GetAccessTokenResponse.java
@@ -1,4 +1,4 @@
-package com.naribackend.api.auth.v1.response;
+package com.naribackend.api.v1.auth.response;
 
 public record GetAccessTokenResponse(
         String accessToken
