+++ src/main/java/com/naribackend/config/WebConfig.java
@@ -1,6 +1,6 @@
 package com.naribackend.config;
 
-import com.naribackend.api.auth.CurrentUserArgumentResolver;
+import com.naribackend.api.CurrentUserArgumentResolver;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
